app/requirements.txt
# DESKTOP VERSION ONLY - 0.9 Beta  - "Orville"

Flask==2.2.2
Flask-SQLAlchemy==3.0.3
requests==2.31.0
MarkupSafe==2.1.3
Werkzeug==2.2.3
mariadb==1.1.8
pytz
python-dotenv
